+++ vector/src/main/java/im/vector/app/features/home/room/detail/RoomDetailViewModel.kt
@@ -57,7 +57,7 @@ import org.commonmark.renderer.html.HtmlRenderer
 import org.matrix.android.sdk.api.MatrixCallback
 import org.matrix.android.sdk.api.MatrixPatterns
 import org.matrix.android.sdk.api.NoOpMatrixCallback
-import org.matrix.android.sdk.api.extensions.tryThis
+import org.matrix.android.sdk.api.extensions.tryOrNull
 import org.matrix.android.sdk.api.query.QueryStringValue
 import org.matrix.android.sdk.api.raw.RawService
 import org.matrix.android.sdk.api.session.Session
@@ -334,7 +334,7 @@ class RoomDetailViewModel @AssistedInject constructor(
             val roomId: String = room.roomId
             val confId = roomId.substring(1, roomId.indexOf(":") - 1) + widgetSessionId.toLowerCase(VectorLocale.applicationLocale)
 
-            val preferredJitsiDomain = tryThis {
+            val preferredJitsiDomain = tryOrNull {
                 rawService.getElementWellknown(session.myUserId)
                         ?.jitsiServer
                         ?.preferredDomain
@@ -988,7 +988,7 @@ class RoomDetailViewModel @AssistedInject constructor(
                 && mxcUrl?.startsWith("content://") ?: false
         val isDownloaded = mxcUrl?.let { session.fileService().isFileInCache(it, action.messageFileContent.mimeType) } ?: false
         if (isLocalSendingFile) {
-            tryThis { Uri.parse(mxcUrl) }?.let {
+            tryOrNull { Uri.parse(mxcUrl) }?.let {
                 _viewEvents.post(RoomDetailViewEvents.OpenFile(
                         action.messageFileContent.mimeType,
                         it,